+++ src/LazyMemberResolver.h
@@ -115,7 +115,7 @@ public:
 		
 		double hit_rate() const {
 			size_t total = cache_hits + cache_misses;
-			return total > 0 ? static_cast<double>(cache_hits) / total : 0.0;
+			return total > 0 ? static_cast<double>(cache_hits) / static_cast<double>(total) : 0.0;
 		}
 	};
